--- TPRB/render/__init__.py
@@ -1,5 +1,0 @@
-from TPRB.render.title import setTitle
-from TPRCBL.parser.pPage import ParsedPage
-
-def render(page: ParsedPage) -> None:
-    setTitle(page.title)
