--- TPRB/render/title.py
@@ -1,4 +1,0 @@
-from os import system
-
-def setTitle(title: str):
-    system('title ', title)
