--- TPRCBL/parser/file.py
@@ -1,5 +1,0 @@
-from TPRCBL.parser.page import Page
-
-def file(path: str) -> Page :
-    with open(path, 'r') as f:
-        return Page(dict(f.read()))
