--- TPRCBL/parser/pPage.py
@@ -1,6 +1,0 @@
-from TPRCBL.elements import Element
-
-class ParsedPage:
-    def __init__(self, title: str = 'Page', elenents: list[Element] = []) -> None:
-        self.title = title
-        self.elements = elenents
